fix: add missing await, move oidc functions
